[refac] Add member withdrawal functionality to MemberController
